CongressDataCollectorPy/requirements.txt
aiohttp==3.8.4
azure-cosmos==4.3.1
python-dotenv==1.0.0
